sct_utils: add mkdir(), rm(), rmtree(), copy(), mv()
They can print equivalent shell commands.

+++ scripts/sct_utils.py
@@ -352,11 +352,44 @@ def display_viewer_syntax(files, colormaps=[], minmax=[], opacities=[], mode='',
         printv(cmd + '\n', verbose=1, type='info')
 
 
-def copy(src, dst):
-    """Copy src to dst.
-    If src and dst are the same files, don't crash.
+def mkdir(path, verbose=1):
+    """Create a folder, like os.mkdir
     """
     try:
+        printv("mkdir %s" % (path), verbose=verbose)
+        os.mkdir(path)
+    except Exception as e:
+        raise
+
+def rm(path, verbose=1):
+    """Remove a file, almost like os.remove
+    """
+    try:
+        printv("rm %s" % (path), verbose=verbose)
+        os.remove(path)
+    except Exception as e:
+        raise
+
+def mv(src, dst, verbose=1):
+    """Move a file from src to dst, almost like os.rename
+    """
+    try:
+        printv("mv %s %s" % (src, dst), verbose=verbose)
+        os.rename(src, dst)
+    except Exception as e:
+        raise
+
+def copy(src, dst, verbose=1):
+    """Copy src to dst, almost like shutil.copy
+    If src and dst are the same files, don't crash.
+    """
+    if not os.path.isfile(src):
+        folder = os.path.dirname(src)
+        contents = os.listdir(folder)
+        raise Exception("Couldn't find %s in %s (contents: %s)" \
+         % (os.path.basename(src), folder, contents))
+    try:
+        printv("cp %s %s" % (src, dst), verbose=verbose)
         shutil.copy(src, dst)
     except Exception as e:
         if sys.hexversion < 0x03000000:
@@ -365,6 +398,15 @@ def copy(src, dst):
         else:
             if isinstance(e, shutil.SameFileError):
                 return
+        raise # Must be another error
+
+def rmtree(folder, verbose=1):
+    """Recursively remove folder, almost like shutil.rmtree
+    """
+    try:
+        printv("rm -rf %s" % (folder), verbose=verbose)
+        shutil.rmtree(folder, ignore_errors=True)
+    except Exception as e:
         raise # Must be another error
 
 
@@ -771,7 +813,7 @@ class TempFolder(object):
 
     def cleanup(self):
         """Remove the created folder and its contents."""
-        shutil.rmtree(self.path_tmp, ignore_errors=True)
+        rmtree(self.path_tmp)
 
 
 #=======================================================================================================================
